refactor: script message handlers → managing class level
