+++ ObservableWebView/Package/ObservableWebViewManager.swift
@@ -95,14 +95,25 @@ extension ObservableWebViewManager {
 }
 
 extension ObservableWebViewManager {
-  private struct Delay {
-    /// Perform an action after a set amount of seconds.
-    static func by(_ seconds: Double, closure: @escaping () -> Void) {
-      Timer.scheduledTimer(withTimeInterval: seconds, repeats: false) { _ in
-        closure()
-      }
-    }
+  func addScriptMessageHandler(_ handler: ScriptMessageHandler, forName name: String) {
+    webView.configuration.userContentController.add(ObservableWebViewScriptMessageProxy(handler: handler), name: name)
+    scriptMessageHandlers[name] = handler
+  }
+  
+  func removeScriptMessageHandler(forName name: String) {
+    webView.configuration.userContentController.removeScriptMessageHandler(forName: name)
+    scriptMessageHandlers.removeValue(forKey: name)
   }
 }
 
-
+private class ObservableWebViewScriptMessageProxy: NSObject, WKScriptMessageHandler {
+  weak var handler: ScriptMessageHandler?
+  
+  init(handler: ScriptMessageHandler) {
+    self.handler = handler
+  }
+  
+  func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
+    handler?.didReceiveScriptMessage(message)
+  }
+}
